// src/app/admin/admin-shared/admin.interfaces.ts
export interface Reports {
  ACCT_NO: string;
  ACCT_TRFF_ID: number;
  ACTIVE: number;
  ACTUAL_OR_ESTIMATE: string | number | boolean;
  ADDRESS: string;
  ADULT_GROUNDS_ADMISSION: string | number | boolean;
  AGREE_REPT_NO: string | number | boolean;
  AMOUNT_2: string | number | boolean;
  AMT: number;
  AMT_TO_BILL: number;
  AMT_TO_BILL_FLAG: string | number | boolean;
  AM_AMT: string | number | boolean;
  ANNUAL_BUDGET: string | number | boolean;
  ATTENDANCE: number;
  AZURE_ID: string;
  BUSINESS_NAME: string;
  CDN_FLAG: string | number | boolean;
  CITY: string;
  CLOSE_DATE: string | number | boolean;
  COMPANY: string;
  CONCERT_REVENUE: string | number | boolean;
  CREATED_AT: string;
  DANCE_EVENTS: string | number | boolean;
  DANCE_FLAG: string | number | boolean;
  DATE_OF_EVENT: string;
  DAYS_OPRN: string | number | boolean;
  DAYS_PER_WEEK: string | number | boolean;
  DAYS_PER_WEEK_2: string | number | boolean;
  DESCR: string | number | boolean;
  DUE_DATE: string | number | boolean;
  EMAIL: string;
  EMAIL_VERIFIED: number;
  EST_OR_ACT: string;
  EVENT_CITY: string;
  EVENT_EMAIL: string;
  EVENT_MAIL_ADDR: string;
  EVENT_NAME: string;
  EVENT_PROV_POSTAL: string;
  EVENT_WEB: string | number | boolean;
  FASHION_SHOW: string | number | boolean;
  FAX: string;
  FEES_PAID_TO_PERFORMERS: string | number | boolean;
  FEE_A: string | number | boolean;
  FEE_B: string | number | boolean;
  FIRST_NAME: string;
  FLAG_RPT_IN_PROGRESS: string;
  FM_AMT: string | number | boolean;
  FREE_FLAG: string;
  GROSS_RECEIPTS: number;
  GROSS_TICKET_SALES: string | number | boolean;
  GST_AMT: number;
  HST_AMT: number;
  INTEREST_AMT: string | number | boolean;
  INVOICE_NO: string | number | boolean;
  INV_REQ: string | number | boolean;
  LANG: string;
  LAST_NAME: string;
  LICENSE_FEE_PER_CONCERT: number;
  LIVE_CONCERTS: string | number | boolean;
  LOGIN_ID: string;
  MIN: number;
  MONTHS_IN_OPRN: string | number | boolean;
  MONTHS_IN_OPRN_2: string | number | boolean;
  MONTHS_REPORTED: number;
  NAME_OF_EVENT: string | number | boolean;
  NON_DANCE_EVENTS: string | number | boolean;
  NON_MTA_AMT: string | number | boolean;
  NO_OF_EVENTS: string | number | boolean;
  NO_OF_UNITS: string | number | boolean;
  NO_OF_UNITS_2: string | number | boolean;
  NUMBER_OF_CONCERTS: string | number | boolean;
  OPEN_DATE: string | number | boolean;
  OP_CODE: string;
  PERCENT: number;
  PERFORMER: string;
  PERIOD_REPORTED: string | number | boolean;
  PHONE: string;
  POSTAL_CODE: string;
  PROGRAM: string | number | boolean;
  PROMOTER: string;
  PROMOTER_ADDRESS: string;
  PROMOTER_CITY: string;
  PROMOTER_POSTAL: string;
  PROMOTER_PROVINCE: string;
  PROVINCE: string;
  PST_AMT: number;
  QUARTER: string | number | boolean;
  RATE: number;
  RATIO_AUD_PG_IMP: string | number | boolean;
  RATIO_NONCDN_PG_IMP: string | number | boolean;
  REF_NO: string | number | boolean;
  REPLACED_REPORT_NO: string | number | boolean;
  REPORTING_FREQ: string;
  REPORT_DETAIL_SEQ: number;
  REPORT_ID: number;
  REPORT_NO: number;
  REQUEST_NO: number;
  RETRO_INT_TO_BILL: string | number | boolean;
  REVENUE: string | number | boolean;
  REV_DATE: string | number | boolean;
  ROLE: number;
  ROOM_CAPACITY: string | number | boolean;
  ROOM_NAME: string;
  RPTDTL_SYSROWID: string | number | boolean;
  SEASONAL_IND: string | number | boolean;
  SEASON_BROCHURE: string | number | boolean;
  SEQ_NO: number;
  SIGNER_NAME: string;
  SIGNER_TITLE: string;
  SPORT_TYPE: string | number | boolean;
  START_DATE: string;
  START_DATE_TARIFF: string;
  STOP_DATE: string;
  SUPPLIER_ADDR: string | number | boolean;
  SUPPLIER_CITY: string | number | boolean;
  SUPPLIER_DATE: string | number | boolean;
  SUPPLIER_NAME: string | number | boolean;
  SUPPLIER_PHONE: string | number | boolean;
  SUPPLIER_PROV_POSTAL: string | number | boolean;
  TRANS_AMT: number;
  TRANS_CMNT: string | number | boolean;
  TRANS_CODE: string;
  TRANS_DATE: string;
  TRANS_RETRO_INT: string | number | boolean;
  TRANS_SOURCE: string;
  TRANS_TYPE: string;
  TRFF_DESC_ENG: string;
  TRFF_DESC_FR: string | number | boolean;
  TRFF_GROUP: number;
  TRFF_NAME_ENG: string;
  TRFF_NAME_FR: string;
  TRFF_NO: string;
  TYPE_OF_UNIT: string | number | boolean;
  TYPE_OF_UNIT_2: string | number | boolean;
  USER_ID: number;
  VALID_MONTH: string | number | boolean;
  VALID_YEAR: number;
  VENUE: string;
  VENUE_ADDRESS: string | number | boolean;
  VENUE_CITY: string | number | boolean;
  VENUE_POSTAL: string | number | boolean;
  VENUE_PROVINCE: string | number | boolean;
  WEEKS_IN_OPRN: string | number | boolean;
}